requirements.txt
googlemaps==3.1.3
pyyaml==5.1.2
pandas==1.5.1